test1

+++ file_encryptor.c
@@ -203,7 +203,7 @@ static CpaStatus cipherPerformOp(CpaInstanceHandle cyInstHandle,
     int i;
     char * srcTemp = NULL;
     printf("%d  %u\n",numBuffers-2,numBuffers-2);
-    for(i=0;i<=(int)numBuffers-2;i++){
+    for(i=0;i<=numBuffers-2;i++){
         srcTemp = src + (1024*1024*i);
         pSrcBuffer[i] = NULL;
         printf("%d Temp: %u  size: %d\n",i,srcTemp, (srcTemp-src)/sizeof(char));
